tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
Pillow==10.1.0
fastapi==0.104.1
httpx==0.25.2
